pyproject.toml
[tool.poetry]
name = "cavity"
version = "0.1.0"
description = "To calculate the cavity formation energy"
authors = ["LRG-UCSD"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.22.3"
MDAnalysis = "^2.7.0"
pyyaml = "^6.0.1"
matplotlib = "^3.4.3"
pandas = "^2.2.2"
ase = "^3.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
